+++ Dracula/docs/release-notes.md
@@ -6,6 +6,7 @@
 
 * Changed the Aura of Darkness card.
 * Changed the Bat Form B card.
+* Changed the Drain Essence card.
 * Fixed mismatched button hitboxes in the Blood Tap UI.
 * Improved wording of the Echolocation card.
 * Improved wording of the O Type ship artifact.